AccessibilityAnnotation Prop implementation final changes
