+++ vnext/Microsoft.ReactNative/Fabric/platform/react/renderer/components/view/HostPlatformViewProps.h
@@ -7,6 +7,7 @@
 #include <react/renderer/core/PropsParserContext.h>
 #include "KeyEvent.h"
 #include "WindowsViewEvents.h"
+#include "AccessibilityAnnotation.h"
 
 namespace facebook::react {
 class HostPlatformViewProps : public BaseViewProps {
