+++ vnext/src-win/Libraries/Components/View/ViewAccessibility.d.ts
@@ -13,7 +13,8 @@ import {NativeSyntheticEvent} from '../../Types/CoreEventTypes';
  * @see https://reactnative.dev/docs/accessibility#accessibility-properties
  */
 export interface AccessibilityProps
-  extends AccessibilityPropsAndroid,
+  extends AccessibilityPropsWindows,
+    AccessibilityPropsAndroid,
     AccessibilityPropsIOS {
   /**
    * When true, indicates that the view is an accessibility element.
@@ -120,15 +121,6 @@ export interface AccessibilityProps
    * Indicates to accessibility services to treat UI component like a specific role.
    */
   role?: Role | undefined;
-
-  /**
-   * Tells a person using a screen reader what kind of annotation they
-   * have selected. If available, it will also tell a person the author of the annotation and
-   * the date and time the annotation was posted.
-   *
-   * Note: If typeID is 'Unknown', a typeName must be provided.
-   */
-  accessibilityAnnotation?: AccessibilityAnnotationInfo; //Windows
 }
 
 export type AccessibilityActionInfo = Readonly<{
@@ -296,12 +288,22 @@ export type AnnotationType =
 
 // [Windows]
 export type AccessibilityAnnotationInfo = Readonly<{
-  typeID: AnnotationType;
-  typeName?: string;
-  author?: string;
-  dateTime?: string;
-  target?: string;
+  typeID: AnnotationType | undefined;
+  typeName?: string | undefined;
+  author?: string |undefined;
+  dateTime?: string | undefined;
+  target?: string |undefined;
 }>;
+export interface AccessibilityPropsWindows{
+  /**
+   * Tells a person using a screen reader what kind of annotation they
+   * have selected. If available, it will also tell a person the author of the annotation and
+   * the date and time the annotation was posted.
+   *
+   * Note: If typeID is 'Unknown', a typeName must be provided.
+   */
+  accessibilityAnnotation?: AccessibilityAnnotationInfo; //Windows
+}
 
 export interface AccessibilityPropsAndroid {
   /**
